feat: Thankyou header & footer moved to shared component location
