+++ support-frontend/assets/components/thankYou/headerFooter/heading.tsx
