+++ support-frontend/assets/components/thankYou/headerFooter/thankYouFooter.tsx
