+++ support-frontend/assets/components/thankYou/headerFooter/thankYouHeader.tsx
